pyproject.toml
[project]
name = "newonepizzeria"
version = "0.1.0"
description = "OnePizzeria Chatbot"
authors = [
    { name = "Diego", email = "diego@example.com" }
]
dependencies = [
    "fastapi",
    "uvicorn",
    "python-telegram-bot>=20.8",
    "langchain",
    "langchain-openai",
    "langchain-core",
    "langgraph",
    "supabase",
    "python-dotenv",
    "pydantic"
]
requires-python = ">=3.10"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
